Feat: 챌린지 상세 정보 API 구현 / QueryDsl 적용

See merge request s08-ai-image-sub2/S08P22A501!55

+++ BE/src/main/java/com/choom/domain/mydance/dto/AddMyDanceResponseDto.java
@@ -12,7 +12,7 @@ public class AddMyDanceResponseDto {
     private int score;
     private String matchRate;
     private String videoPath;
-    private double videoLength;
+    private int videoLength;
     private String title;
 
     @Builder
